% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{mme-package}
\alias{mme}
\alias{mme-package}
\title{Multinomial Mixed Effects Models}
\description{
The mme package implements three multinomial area level mixed effects models for
small area estimation. The first model (Model 1) is based on the area level multinomial
mixed model with independent random effects for each category of the response
variable (Lopez-Vizcaino et al, 2013). The second model (Model 2) takes advantage
from the availability of survey data from different time periods and uses
a multinomial mixed model with independent random effects for each category of the
response variable and with independent time and domain random effects. The
third model (Model 3) is similar to the second one, but with correlated time random effects. To fit the models, we combine the penalized
quasi-likelihood (PQL) method, introduced by Breslow and Clayton (1993) for estimating and predicting th fixed
and random effects, with the residual maximum likelihood (REML) method for estimating the variance components.
In all models the package use two approaches to estimate the mean square
error (MSE), first through an analytical expression and second by bootstrap techniques.
}
\references{
Lopez-Vizcaino, ME, Lombardia, MJ and Morales, D (2013).
Multinomial-based small area estimation of labour force indicators.
Statistical Modelling, 13 ,153-178.

Lopez-Vizcaino, ME, Lombardia, MJ and Morales, D (2013).
Small area estimation of labour force indicator under a multinomial mixed model
with correlated time and area effects. Submitted for review.

Breslow, N, Clayton, D (1993).
Aproximate inference in generalized linear mixed models.
Journal of the American Statistical Association, 88, 9-25.
}

