% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MME.R
\name{MME_tune}
\alias{MME_tune}
\title{Fit mixture of Erlangs}
\usage{
MME_tune(lower, upper = lower, trunclower = rep(0, ncol(as.matrix(lower))),
  truncupper = rep(Inf, ncol(as.matrix(lower))), M = 10, s = 100,
  nCores = detectCores(), criterium = "BIC", eps = 0.001,
  beta_tol = 10^(-5), print = TRUE, file = "log.txt", max_iter = Inf)
}
\arguments{
\item{lower}{A matrix specifying the lower censoring points, observations in
rows, dimensions in coloms.}

\item{upper}{A matrix specifying the upper censoring points, observations in
rows, dimensions in coloms.}

\item{trunclower}{A vector specifying the lower truncation points in each
dimension.}

\item{truncupper}{A vector specifying the upper truncation points in each
dimension.}

\item{M}{A vector of values for the tuning parameter M.}

\item{s}{A vector of values for the tuning parameter s (the spread).}

\item{nCores}{Number of cores available for parallel computation.}

\item{criterium}{Character vector specifying information criterium to use,
either "AIC" or "BIC".}

\item{eps}{Covergence threshold used in the EM algorithm.}

\item{beta_tol}{Threshold for the mixing weights below which the
corresponding shape parameter vector is considered neglectable.}

\item{print}{Bool: print intermediate results, either TRUE or FALSE.}

\item{file}{If print is TRUE, specify file name.}

\item{max_iter}{Maximum number of iterations in a single EM algorithm.}
}
\value{
A list containing:
\describe{
   \item{best_model}{The final MME, judged to be the best according to the criterium used. Value is a list.}
   \item{performances}{A matrix summarizing the performance of the different fitted MME models (M, s, criterium, R).}
   \item{all_model}{A list containing all fitted MME models.}
}
}
\description{
Fits univariate and multivariate mixtures of Erlang distributions to possibly
censored and/or truncated data. The censoring and/or truncation can be left,
right or interval.
}
\examples{
\dontrun{
library(MASS)
MME_tune(lower = geyser, upper = geyser, trunclower = c(0, 0),
truncupper = c(Inf, Inf), M = c(5, 10, 20), s = c(seq(10, 100, 10), 200),
nCores = detectCores(), criterium = "BIC", eps = 1e-03,
 beta_tol = 10^(-5), print=TRUE, file="log.txt", max_iter = Inf)
}
}

